requirements.txt
django==1.11.1
djangorestframework==3.6.2
Pillow==4.1.1
django-rest-swagger==2.1.2
